tests/ambient-light/ambient-light_browsing_context.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: browsing context</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Xiuqi Jiang" href="mailto:xiuqix.jiang@intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<link rel="help" href="https://w3c.github.io/sensors/releases/WD-generic-sensor-20160830.html#browsing-context">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<div style="display:none">
<iframe width=100 height=100 src="support-iframe.html" id="iframe">
</iframe>
</div>
<script>

async_test(function (t) {
  document.getElementById('iframe').onload = t.step_func_done(function(event) {
    var iframe = document.getElementById('iframe').contentWindow;
    var content = iframe.document.getElementById('iframe').value;
    assert_equals(content, "null");
  });
}, "the sensor readings can not be fired within iframes");

test(function() {
  try {
    var str_sensor_reading;
    let sensor = new AmbientLightSensor();

    sensor.start();

    var win = window.open('support-window-open.html', '_blank');
    str_sensor_reading = String(sensor.reading)

    win.close();
    sensor.stop();
    assert_equals(str_sensor_reading, "null");
  }
  catch (ex) {
    assert_unreached('threw an unexpected exception "' + ex.name + '"');
  }
}, "the sensor readings must only be available in the top-level browsing context");

</script>
